driver::BaseLedDriver: getLedIndex() should take an int
Because we do not necessarily have access to `KeyAddr` at that point, lets just
take an int. Nothing is using this method directly, all calls are made through
`Device::getLedIndex()`, which _does_ have access to `KeyAddr`. As such,
`Device::getLedIndex()` will continue taking a `KeyAddr`, and will convert it to
an int when calling the LED driver.

This makes it possible to define the LED description before defining `KeyAddr`
itself.

Also updated the Model01 LED driver to match the new interface.

+++ src/kaleidoscope/Device.h
@@ -176,7 +176,7 @@ class Device {
   * LEDs there.
   */
   int8_t getLedIndex(KeyAddr key_addr) {
-    return leds_.getLedIndex(key_addr);
+    return leds_.getLedIndex(key_addr.toInt());
   }
   /**
    * Returns the index of the LED at a given row & column.
